Adiciona tradução para os elementos
